deploy-test

+++ app/controllers/favorites_controller.rb
@@ -1,16 +1,19 @@
 class FavoritesController < ApplicationController
   def create
-    post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: post.id)
-    favorite.save
-    redirect_back(fallback_location: root_path)
+    @post = Post.find(params[:post_id])
+    if @post.user != current_user
+      favorite = current_user.favorites.new(post_id: @post.id)
+      favorite.save
+      @post.create_notification_favorite!(current_user)
+    end
   end
 
   def destroy
-    post = Post.find(params[:post_id])
-    favorite = current_user.favorites.find_by(post_id: post.id)
-    favorite.destroy
-    redirect_back(fallback_location: root_path)
+    @post = Post.find(params[:post_id])
+    if @post.user != current_user
+      favorite = current_user.favorites.find_by(post_id: @post.id)
+      favorite.destroy
+    end
   end
 
 end
